# app/api-docs/swagger.yaml
openapi: 3.0.3
info:
  title: Figma AI Ticket Generator API
  description: |
    🎯 **Unified Documentation Generation API**
    
    Generate JIRA tickets, Wiki pages, Code documentation, and more from Figma designs using AI.
    
    ## 🚀 **Key Features:**
    - **Unified Generation**: Single endpoint for all documentation formats
    - **Simplified Strategies**: AI-Powered (primary), Emergency (fallback)
    - **Format Support**: JIRA (active), Wiki/Code/Markdown (planned)
    - **Smart Auto-Detection**: Automatically selects optimal approach based on input
    - **Legacy Compatibility**: Backward compatible with deprecation warnings
    
    ## 🎯 **Primary Endpoint:**
    Use `POST /api/generate` for all documentation generation needs.
    
    ## 📋 **Migration Guide:**
    - `POST /api/generate-ticket` → `POST /api/generate` with `format: "jira"` and `strategy: "ai-powered"`
    - `POST /api/generate-ai-ticket-direct` → `POST /api/generate` with `format: "jira"` and `strategy: "ai-powered"`
    
  version: 2.0.0
  contact:
    name: Figma AI Support
    url: https://github.com/figma-ai/ticket-generator
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development Server
  - url: https://api.figma-ai.com
    description: Production Server
...
paths:
...
  /plugin/comprehensive-context:
    post:
      tags: [plugin]
      summary: 🚀 Comprehensive Context Collection (Plugin API)
      description: |
        **🔌 Figma Plugin Modular API System**
        
        This documents the **client-side plugin API** that provides comprehensive context collection 
        directly from Figma using our modular system. This API operates via `postMessage` 
        communication between the Figma plugin UI and the plugin code.
        
        ### 🏗️ Architecture:
        ```
        Figma UI "Preview Context" Button Click 
          ↓
        collectComprehensiveContext() [Client-side]
          ↓
        9x Modular Collection Functions (Parallel)
          ↓
        postMessage() to Figma Plugin
          ↓
        6x Plugin Message Handlers
          ↓
        Direct Figma API Extraction
          ↓
        Structured Data Storage
        ```
        
        ### 📦 **9 Collection Modules:**
        1. **`figmaSelection`** - Core selection data with hierarchy
        2. **`figmaComponents`** - Component analysis and variants
        3. **`figmaStyles`** - Design tokens extraction
        4. **`figmaInteractions`** - Prototype interactions  
        5. **`figmaConstraints`** - Auto-layout & responsive data
        6. **`figmaEffects`** - Shadows, blurs, advanced styling
        7. **`screenshot`** - High-quality image capture
        8. **`accessibility`** - A11y analysis from collected data
        9. **`codeGeneration`** - Implementation readiness specs
        
        ### 🎯 **Plugin Message Handlers:**
        - `get-comprehensive-selection` → `handleGetComprehensiveSelection()`
        - `analyze-components` → `handleAnalyzeComponents()`
        - `extract-design-tokens` → `handleExtractDesignTokens()`
        - `analyze-interactions` → `handleAnalyzeInteractions()`
        - `analyze-constraints` → `handleAnalyzeConstraints()`
        - `analyze-effects` → `handleAnalyzeEffects()`
        
        ### 💾 **Data Storage:**
        ```javascript
        window.comprehensiveContextData = {
          metadata: { version: '3.0.0', collectedAt: '...', source: '...' },
          figmaData: { selection: [...], components: [...], interactions: [...] },
          designTokens: { colors: [...], typography: [...], spacing: [...] },
          technicalSpecs: { constraints: {...}, effects: {...} },
          accessibility: { contrastRatios: [...], wcagCompliance: '...' },
          codeGeneration: { framework: '...', implementation: {...} },
          screenshot: { url: '...', metadata: {...} }
        }
        ```
        
        ### ⚡ **Trigger Mechanism:**
        **Only triggers on "🔍 Preview Context & Generate" button click**
        
        **No HTTP requests needed** - all data comes directly from Figma API via plugin context.

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [techStack]
              properties:
                techStack:
                  type: string
                  description: Target technology stack
                  example: "React with TypeScript"
                documentType:
                  type: string
                  enum: [story, task, bug, epic, component]
                  default: story
                  description: Type of documentation to generate
                enableModules:
                  type: object
                  description: Module configuration - enable/disable specific collection modules
                  properties:
                    figmaSelection:
                      type: boolean
                      default: true
                      description: Core selection data with hierarchy
                    figmaComponents:
                      type: boolean
                      default: true
                      description: Component analysis and variants
                    figmaStyles:
                      type: boolean
                      default: true
                      description: Design tokens extraction
                    figmaInteractions:
                      type: boolean
                      default: true
                      description: Prototype interactions analysis
                    figmaConstraints:
                      type: boolean
                      default: true
                      description: Auto-layout & responsive constraints
                    figmaEffects:
                      type: boolean
                      default: true
                      description: Shadows, blurs, advanced styling
                    screenshot:
                      type: boolean
                      default: true
                      description: High-quality screenshot capture
                    accessibility:
                      type: boolean
                      default: true
                      description: Accessibility analysis from data
                    codeGeneration:
                      type: boolean
                      default: true
                      description: Code generation readiness assessment
            examples:
              full_collection:
                summary: Full Comprehensive Collection
                description: Collect all 9 modules for maximum context
                value:
                  techStack: "React with TypeScript"
                  documentType: "component"
                  enableModules:
                    figmaSelection: true
                    figmaComponents: true
                    figmaStyles: true
                    figmaInteractions: true
                    figmaConstraints: true
                    figmaEffects: true
                    screenshot: true
                    accessibility: true
                    codeGeneration: true
              selective_collection:
                summary: Selective Module Collection
                description: Only collect essential modules for speed
                value:
                  techStack: "Vue.js"
                  documentType: "story"
                  enableModules:
                    figmaSelection: true
                    screenshot: true
                    designTokens: true
                    accessibility: false
                    codeGeneration: false
              minimal_collection:
                summary: Minimal Data Collection
                description: Basic selection and screenshot only
                value:
                  techStack: "Angular"
                  documentType: "task"
                  enableModules:
                    figmaSelection: true
                    screenshot: true
                    # All other modules disabled for speed

      responses:
        '200':
          description: Comprehensive context collection completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Comprehensive context collected successfully"
                  data:
                    $ref: '#/components/schemas/ComprehensiveContextData'
                  performance:
                    type: object
                    properties:
                      totalModules:
                        type: integer
                        example: 9
                      completedModules:
                        type: integer
                        example: 9
                      collectionTime:
                        type: integer
                        description: Total collection time in milliseconds
                        example: 3420
                      averageModuleTime:
                        type: integer
                        description: Average time per module in milliseconds
                        example: 380
                      dataPoints:
                        type: integer
                        description: Total number of data points collected
                        example: 847
              examples:
                successful_collection:
                  summary: Successful Comprehensive Collection
                  value:
                    success: true
                    message: "Comprehensive context collected from 9 modules"
                    data:
                      metadata:
                        collectedAt: "2025-11-07T15:30:00.000Z"
                        source: "modular-comprehensive-collection"
                        version: "3.0.0"
                        enabledModules: 
                          figmaSelection: true
                          figmaComponents: true
                          figmaStyles: true
                          screenshot: true
                      figmaData:
                        selection: 
                          - id: "123:456"
                            name: "Login Button"
                            type: "INSTANCE"
                            hierarchy: { layers: 3, componentCount: 1 }
                        components:
                          - id: "123:456"
                            name: "Button Component"
                            type: "instance"
                            masterComponent: { id: "789:012", name: "Base Button" }
                      designTokens:
                        colors: ["#007AFF", "#34C759", "#FF3B30"]
                        typography: ["Inter-16px-Medium", "Inter-14px-Regular"]
                        spacing: [8, 16, 24, 32]
                        borderRadius: [4, 8, 12]
                      screenshot:
                        url: "data:image/png;base64,iVBORw0KGgoAAAANS..."
                        metadata: { captureTime: "2025-11-07T15:30:00.000Z" }
                    performance:
                      totalModules: 9
                      completedModules: 9
                      collectionTime: 3420
                      averageModuleTime: 380
                      dataPoints: 847

        '400':
          description: Invalid module configuration or missing required fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Missing required field: techStack"
                  modules:
                    type: object
                    properties:
                      attempted:
                        type: array
                        items:
                          type: string
                      failed:
                        type: array
                        items:
                          type: string
                      succeeded:
                        type: array
                        items:
                          type: string

        '408':
          description: Module collection timeout
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Module collection timed out after 15 seconds"
                  partialData:
                    $ref: '#/components/schemas/ComprehensiveContextData'
                    description: Partial data from completed modules
                  completedModules:
                    type: array
                    items:
                      type: string
                    example: ["figmaSelection", "screenshot", "designTokens"]
                  failedModules:
                    type: array
                    items:
                      type: string
                    example: ["figmaComponents", "figmaInteractions"]
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (when enabled)
security:
  - {} # No security required for development
  # - ApiKeyAuth: [] # Uncomment for production
...